fixed compiler warning in pcl/io/src/lzf.cpp:192
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4609 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ io/src/lzf.cpp
@@ -189,7 +189,7 @@ pcl::lzfCompress (const void *const in_data, unsigned int in_len,
 
         do
           len++;
-        while (len < maxlen && ref[len] == ip[len]);
+        while (len < (unsigned int)maxlen && ref[len] == ip[len]);
 
         break;
       }
